+++ src/main/java/io/kontur/eventapi/metrics/MetricCollector.java
@@ -1,0 +1,5 @@
+package io.kontur.eventapi.metrics;
+
+public interface MetricCollector {
+    void collect();
+}