+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - index `normalized_observations_source_updated_collected_geography_idx`
 - metrics on SQS queue sizes
 - Gitlab CI test reports
+- re-enrichment job
 
 The following config should be added to all the stages: